Refactor `convert_resources_to_resource_model` to `core.resources`
